// src/app/FeatureModules/Customer/Models/Customer/Customer.ts
export class CustomerModel {
  "customerTypeId": number = -1;
  "customerName": string;
  "defaultEmail": boolean;
  "contact": string;
  "email": string;
  "address1": string;
  "address2": string;
  "address3": string;
  "suburbId": number;
  "state": string;
  "postCode": number;
  "sameAsStreet": boolean;
  "address1Postal": string;
  "address2Postal": string;
  "address3Postal": string;
  "suburbPostal": number;
  "statePostal": string;
  "postCodePostal": number;
  "invoiceMethod": number;
  "createdBy": 'Micheal';
}