// src/app/FeatureModules/Customer/Models/Customer/UpdateCustomer.ts
export class UpdateCustomer {
   "customerId": number;
   "customerTypeId": number;
   "customerName": string;
   "contact": string;
   "defaultEmail": boolean;
   "sameAsStreet": boolean;
   "email": string;
   "address1": string;
   "address2": string;
   "address3": string;
   "suburbId": number;
   "state": string;
   "postCode": number;
   "address1Postal": string;
   "address2Postal": string;
   "address3Postal": string;
   "suburbPostal": string;
   "statePostal": string;
   "postCodePostal": number;
   "invoiceMethod": number;
   "modifiedBy": "Micheal";
}

export class AccountBalance {
   CustomerId: number;
}